# docs/qa/openapi.yaml
...
openapi: 3.0.3
info:
  title: Stitch ATS API
  description: "REST API for Stitch ATS — recruitment, candidate pipeline, interviews, offers, and portals.\n\n**Authentication:** JWT via `POST /api/auth/login`. Pass `Authorization: Bearer <token>` on protected routes.\n\n**Integration key:** Microsoft 365 routes use header `X-Integration-Api-Key`.\n\nSee also: [API_REFERENCE.md](./API_REFERENCE.md) for QA test accounts and workflows."
  version: 1.0.0
  contact:
    name: Stitch ATS
    url: "https://stitch-ats.in"
servers:
  - url: "https://qa.stitch-ats.in"
    description: QA staging (frontend proxy)
  - url: "https://ats-0dtj.onrender.com"
    description: QA API (Render direct)
  - url: "https://stitch-ats.in"
    description: Production (frontend proxy)
  - url: "https://stitch-ats.onrender.com"
    description: Production API (Render direct)
  - url: "http://localhost:3001"
    description: Local dev API
...
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: List users
      description: "**Roles:** ADMIN, HR_HEAD, HR_MANAGER, RECRUITER, TEAM_LEAD, HIRING_MANAGER, INTERVIEWER"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: User array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Users
      summary: Create staff user
      description: "**Roles:** SUPER_ADMIN, ADMIN"
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        201:
          description: Created user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  inviteEmailSent:
                    type: boolean
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        409:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/business-requirements:
    get:
      tags:
        - Business Requirements
      summary: List business requirements
      description: "**Roles:** ACCOUNT_MANAGER, HIRING_MANAGER, ADMIN"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessRequirement"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Business Requirements
      summary: Create business requirement
      description: "**Roles:** ACCOUNT_MANAGER, HIRING_MANAGER, ADMIN"
      security:
        -
          bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        201:
          description: Object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessRequirement"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/requirements:
    get:
      tags:
        - Requirements
      summary: List requirements
      description: "**Roles:** Internal (scoped)"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Requirement"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Requirements
      summary: Create requirement
      description: "Required: title, department, hiringManager, openings, priority, client, primarySkills, secondarySkills"
      security:
        -
          bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        201:
          description: Object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requirement"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/candidates:
    get:
      tags:
        - Candidates
      summary: List candidates
      description: "**Roles:** Internal (INTERVIEWER blocked)"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Candidates
      summary: Create candidate
      description: "**Roles:** Staff mutate"
      security:
        -
          bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              additionalProperties: true
      responses:
        201:
          description: Object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/interviews:
    get:
      tags:
        - Interviews
      summary: List interviews
      description: "**Roles:** Internal (scoped)"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Interviews
      summary: Schedule interview
      description: "**Roles:** Interview schedulers"
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - requirementId
                - candidateId
                - planStageId
                - scheduledAt
                - interviewerIds
              properties:
                requirementId:
                  type: string
                candidateId:
                  type: string
                planStageId:
                  type: string
                scheduledAt:
                  type: string
                  format: date-time
                interviewerIds:
                  type: array
                  items:
                    type: string
                  minItems: 1
                type:
                  type: string
                status:
                  $ref: "#/components/schemas/InterviewStatus"
                meetingLink:
                  type: string
                duration:
                  type: integer
                location:
                  type: string
                description:
                  type: string
      responses:
        201:
          description: Object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/offers:
    get:
      tags:
        - Offers
      summary: List offers
      description: "**Roles:** Offer roles (SUPER_ADMIN, ADMIN, HR_HEAD, HR_MANAGER, RECRUITER, TEAM_LEAD)"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Offers
      summary: Create offer draft
      description: "**Roles:** Offer roles (SUPER_ADMIN, ADMIN, HR_HEAD, HR_MANAGER, RECRUITER, TEAM_LEAD)"
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - candidateId
                - requirementId
              properties:
                candidateId:
                  type: string
                requirementId:
                  type: string
                annualCtc:
                  type: number
                letterMeta:
                  type: object
                approvalChain:
                  type: object
      responses:
        201:
          description: Object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/feedback:
    post:
      tags:
        - Feedback
      summary: Submit feedback
      description: "**Roles:** Internal"
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - interviewId
                - candidateId
              properties:
                interviewId:
                  type: string
                candidateId:
                  type: string
                interviewerId:
                  type: string
                rating:
                  type: number
                technicalRating:
                  type: number
                communicationRating:
                  type: number
                comments:
                  type: string
                recommendation:
                  $ref: "#/components/schemas/FeedbackRecommendation"
                formData:
                  type: object
      responses:
        201:
          description: Object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/activity-logs:
    get:
      tags:
        - Activity Logs
      summary: Recent activity logs
      description: "**Roles:** Internal"
      security:
        -
          bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        200:
          description: Array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Activity Logs
      summary: Create activity log
      description: "**Roles:** Internal"
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - entityType
                - entityId
                - action
              properties:
                entityType:
                  type: string
                entityId:
                  type: string
                action:
                  type: string
                details:
                  type: object
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/search:
    get:
      tags:
        - Search
      summary: Global search
      description: "**Roles:** Internal"
      security:
        -
          bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  candidates:
                    type: array
                    items:
                      $ref: "#/components/schemas/Candidate"
                  requirements:
                    type: array
                    items:
                      $ref: "#/components/schemas/Requirement"
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  interviews:
                    type: array
                    items:
                      $ref: "#/components/schemas/Interview"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/role-access:
    get:
      tags:
        - Role Access
      summary: Full access matrix
      description: "**Roles:** SUPER_ADMIN"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Access config
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/interview-panels:
    get:
      tags:
        - Interview Panels
      summary: List panel levels
      description: "**Roles:** Internal"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Panel levels
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /api/auth/login (~7 day lifetime)
    integrationApiKey:
      type: apiKey
      in: header
      name: X-Integration-Api-Key
      description: M365_INTEGRATION_API_KEY env var
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
        code:
          type: string
      required:
        - error
...
    Candidate:
      type: object
      description: Full candidate record with pipeline fields
      additionalProperties: true
    Requirement:
      type: object
      description: Job requirement / requisition
      additionalProperties: true
    BusinessRequirement:
      type: object
      description: Pre-hiring business requirement with deal stages
      additionalProperties: true
...
    Feedback:
      type: object
      additionalProperties: true
...